requirements.txt
amqp==2.2.2
astroid==1.6.3
billiard==3.5.0.3
cache.py==0.1.3
caching==0.1.dev8
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.5.0
dill==0.2.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-cache-machine==1.0.0
django-cache-manager==0.5.1
django-cors-headers==2.2.0
django-filter==1.1.0
django-model-utils==3.1.2
django-oauth-toolkit==1.1.2
django-toolbelt==0.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
httpie==0.9.9
idna==2.6
image==1.5.24
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
joblib==0.11
kiwisolver==1.0.1
kombu==4.2.0
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
numpy==1.14.3
oauthlib==2.0.7
pandas==0.23.0
pep8==1.7.1
Pillow==5.1.0
pipenv==2018.5.18
psycopg2==2.7.4
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
PyMySQL==0.8.1
pyparsing==2.2.0
pyreadline==2.1
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
rest-social-auth==1.3.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlitedict==1.5.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
wrapt==1.10.11
